network 설정 옮김
