+++ Electo/Electo/ClassifiedPhotoViewController.swift
@@ -143,26 +143,6 @@ class ClassifiedPhotoViewController: UIViewController {
             self?.tableView.reloadData()
         }
     }
-    
-    @IBAction func networkAllowSwitch(_ sender: UISwitch) {
-       print(sender.state)
-        if sender.isOn {
-            let alertController = UIAlertController(title: "", message: "It will use network data", preferredStyle: .alert)
-            let okAction = UIAlertAction(title: "OK", style: .default, handler: { (action) in
-                Constants.dataAllowed = true
-            })
-            let cancelAction = UIAlertAction(title: "cancel", style: .cancel, handler: { (action) in
-                Constants.dataAllowed = false
-            })
-            alertController.addAction(okAction)
-            alertController.addAction(cancelAction)
-            print("on")
-            present(alertController, animated: true, completion: nil)
-        } else {
-            print("off")
-            Constants.dataAllowed = false
-        }
-    }
 }
 
 extension ClassifiedPhotoViewController: UITableViewDelegate {
