+++ Electo/Electo/SettingViewController.swift
@@ -1,0 +1,32 @@
+//
+//  SettingViewController.swift
+//  Electo
+//
+//  Created by Alpaca on 2017. 8. 17..
+//  Copyright © 2017년 RodoPacaGiraffe. All rights reserved.
+//
+
+import UIKit
+
+class SettingViewController: UITableViewController {
+    
+    @IBAction func networkAllowSwitch(_ sender: UISwitch) {
+        print(sender.state)
+        if sender.isOn {
+            let alertController = UIAlertController(title: "", message: "It will use network data", preferredStyle: .alert)
+            let okAction = UIAlertAction(title: "OK", style: .default, handler: { (action) in
+                Constants.dataAllowed = true
+            })
+            let cancelAction = UIAlertAction(title: "cancel", style: .cancel, handler: { (action) in
+                Constants.dataAllowed = false
+            })
+            alertController.addAction(okAction)
+            alertController.addAction(cancelAction)
+            print("on")
+            present(alertController, animated: true, completion: nil)
+        } else {
+            print("off")
+            Constants.dataAllowed = false
+        }
+    }
+}
